phase-helpers.sh: avoid repeated PATH searches for gpatch
Presently, the eapply() function checks whether gpatch is available by
using the hash builtin. If found, it is presumed to be GNU patch(1) and
is selected as the preferred implementation. However, in the unlikely
event that it is missing, the hash builtin will incur a PATH search upon
each occasion that the function is called. That is because the cache of
utility locations can only contain positive entries (found paths by
utility name), not negative entries (unfound paths by utility name).

Address this minor issue by having the "phase-helpers.sh" unit check for
the existence of gpatch at the time that it is sourced, and by having it
declare a patch() function that serves as a wrapper.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>