emerge-webrsync: improve the diagnostic messages of sync_local()
As concerns the sync_local() function, increase the granularity and
utility of the informational messages that are shown. The revised
messages convey the following information:

- which utility is used to sync the repo, be it rsync(1) or tarsync
- the point at which tarball extraction begins, if using rsync(1)
- the location of the repo

Further, refrain from issuing suppositional claims in the event that
tar(1) or tarsync fail. There are various potential modes of failure and
it is a stretch to imply that the tarball be corrupt after undergoing
verification by checksum and - in most cases - by gemato or gpg(1).

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>