emerge-webrsync: parse FETCHCOMMAND only once in fetch_file()
Presently, the fetch_file() function considers the value of the
'FETCHCOMMAND' portage variable and composes a derivative command to be
conveyed to the eval builtin. Further, it does so on each and every
occasion that the function is called. Instead, have it compose the
command just once. This is accomplished by stashing the command in the
newly introduced 'fetchcommand' variable, residing in the global scope.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>