emerge-webrsync: refrain from calling die() before it is declared
Presently, the emerge-webrsync script begins by trying to locate and
store the paths of the emerge(1) and portageq utilities. In the case
that either is missing, a diagnostic message is displayed before
exiting, with the aid of the die() function. However, the function will
not yet have been declared.

Address this issue by avoiding the use of the die() function at that
particular juncture. It was once the case that the function was declared
prior, but the declaration was deliberately repositioned so as to
override the over-engineered and unwieldy function of the same name from
the "isolated-functions.sh" unit.

Fixes: a9b74b15ebdb402b36959c74b53be8110d09e458
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>