emerge-webrsync: don't tolerate an empty definition of FETCHCOMMAND
Presently, the get_fetchcommand() function considers the value of the
'FETCHCOMMAND' portage variable and composes a derivative command to be
conveyed by the fetch_file() function to the eval builtin. However, it
will tolerate a value that is either empty or blank, in which case the
invocation of the eval builtin will successfully do nothing.

Address this issue by having the get_fetchcommand() function check
whether the value of 'FETCHCOMMAND' is composed entirely of zero-or-more
blank characters and returning false in that event. The fetch_file()
function will duly respond by dying.

 * FETCHCOMMAND has been set as an empty or blank string
emerge-webrsync: couldn't parse FETCHCOMMAND

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>