emerge-webrsync: move snapshot freshness check to a distinct function
Presently, the do_snapshot() function composes an array of snapshot
tarballs and an array of mirrors before iterating over the cross product
of both, stopping after encountering a tarball that can be fetched and
validated. There are three modes of validation, which are as follows.

- verifying against the published MD5 checksum
- verifying with gemato or gpg(1)
- verifying that the timestamp falls within an acceptable time range

The latter mode of verification is sufficiently complex as to merit its
own function. Make it so by introducing the is_snapshot_fresh() function
and calling it from do_snapshot().

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>