emerge-webrsync: improve legibility of tests for already-fetched files
Presently, the do_snapshot() function will attempt to make use of a
previously downloaded snapshot tarball in the case that it is found to
exist in the directory specified by the 'DISTDIR' variable, provided
that it is accompanied by its ".md5sum" and ".gpgsig" files. This commit
renders the associated tests more legible.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>