emerge-webrsync: jettison the do_tar() function
GNU tar(1) is perfectly capable of determining whether its input needs
to be decompressed as a bzip2 or xz stream. So is bsdtar(1), provided
that libarchive has been built accordingly; both "bzip2" and "zstd" are
default-enabled by IUSE in app-arch/libarchive. As such, the do_tar()
function serves no purpose. Get rid of it.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>