emerge-webrsync: support bsdtar
Presently, emerge-webrsync is incompatible with bsdtar(1) because the
get_snapshot_timestamp() function specifies the --wildcards option.

Address this issue by only specifying this option in the case that the
tar(1) implementation is found to be that of GNU. The BSD implementation
always treats its operands as globs, provided that either the -t or -x
option has been specified.

It should be noted that this renders emerge-webrsync compatible with the
native tar(1) implementations of macOS, FreeBSD and NetBSD, but not
those of OpenBSD, MirBSD and Solaris. This matters little because the
Gentoo Prefix installation script bootstraps GNU tar. Nonetheless, there
is no harm in improving upon the portability of emerge-webrsync.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>