ebuild.sh: rectify a format string injection
Presently, the "ebuild.sh" utility contains a routine that merges the
values of the 'SANDBOX_DENY', 'SANDBOX_PREDICT', 'SANDBOX_READ' and
'SANDBOX_WRITE' variables with those of their counterparts that are
prefixed with "PORTAGE_". In the course of doing so, it directly injects
arbitrary pathnames into the format string given to an invocation of the
printf builtin. Refrain from doing so.

Fixes: 078abd42ede4b69f618b67c86a698030fe9d3c3b
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>