ebuild.sh: merge sandbox paths without sort -z and elide empty paths
Presently, the "ebuild.sh" utility contains a routine that merges the
colon-separated substrings of the 'SANDBOX_DENY', 'SANDBOX_PREDICT',
'SANDBOX_READ' and 'SANDBOX_WRITE' variables with those of their
counterparts that are prefixed with "PORTAGE_". In the course of doing
so, it executes the sort(1) and tr(1) utilities so as to de-duplicate
the resulting list, effectively producing a set. This routine suffers
from a few minor issues, which are described and addressed herewith.

The sort(1) utility is given the non-standard -z option. I find this to
be distasteful, chiefly because PMS makes no promise that a given
implementation of the utility shall be available. Address this issue by
refraining from using any external utilities whatsoever.

The routine fails to elide empty paths. Address this issue by
consistently disregarding the null string. Consider a scenario in which
the 'SANDBOX_DENY' and 'PORTAGE_SANDBOX_DENY' variables are being
processed, having been set as follows.

          SANDBOX_DENY=/foo::/bar
  PORTAGE_SANDBOX_DENY=/baz

Prior to this commit, the string representing the final set would be
formed as ":/bar:/baz:/foo". Following this commit, it shall instead be
formed as "/baz:/foo:/bar".

It should be noted that the revised code employs the ${param@Q} form of
expansion. Doing so is acceptable, given a target of >=bash-4.4.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>