Define four variables in the PORTAGE_ namespace as lowercase arrays
Presently, the "phase-functions.sh" unit defines the following four
scalar variables.

- PORTAGE_MUTABLE_FILTERED_VARS
- PORTAGE_SAVED_READONLY_VARS
- PORTAGE_READONLY_METADATA
- PORTAGE_READONLY_VARS

Their values are comprised by a whitespace-separated list of variable
names, which are referenced in as many as two ways. Firstly, by the
__filter_readonly_variables() function as the names of variables that
are to be filtered. Secondly, by the "ebuild.sh" utility as the names of
variables for which the readonly attribute should be set.

Given that none of these variables are used in any other capacity, and
that all are acting as mock-arrays, declare them as actual array
variables instead. That way, no word splitting need be performed, be it
with the aid of the read builtin or by way of an unquoted expansion. Two
instances of SC2086 are duly addressed.

Further, rename the variables so as to be in lower case. Given that it
is a widespread convention for environment variables to have names that
are entirely in upper case, this renders the provenance and purpose of
the variables clearer than would otherwise be the case. That is, none of
them are expected to:

- have been defined as an environment variable by the parent process
- be designated for exporting to subprocesses
- be referenced in any other capacity (such as by an ebuild or eclass)

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>