phase-functions.sh: reorder hard-coded array elements to optimise for legibility
Presently, the "phase-functions.sh" unit defines the following four
array variables in the global scope.

- PORTAGE_MUTABLE_FILTERED_VARS
- PORTAGE_SAVED_READONLY_VARS
- PORTAGE_READONLY_METADATA
- PORTAGE_READONLY_VARS

It also defines the following three array variables as local to the
__filter_readonly_variables() function.

- binpkg_untrusted_vars
- filtered_sandbox_vars
- bash_vars

The elements contained by all of these arrays are hard-coded and are
presented in ascending lexicographical order. However, I have found that
sequences of identifiers can be rendered more legible by being in
descending order of length. At the same time, maintaining an
alphabetical order makes it easier to identify a particular element
while perusing longer sequences. With both of these considerations in
mind, I devised a simple algorithm that sorts in a manner threefold:

1) by ascending lexicographical order of that which matches ^_*[[:alpha:]]
2) by descending order or element length, in characters
3) by asending lexicographical order

I found the results to be quite satisfactory. As such, this commit
reorders the elements of the aforementioned array variables in
accordance with this algorithm. The improvement in legibility is most
notable in the case of the 'portage_readonly_vars' variable. For
reference, a perl implementation of this algorithm is shown beneath.

#!/usr/bin/perl
while (readline *STDIN) {
        push @words, m/\S+/g;
}
@words = sort {
        ($c1, $c2) = map { m/^(_*\p{PosixAlpha})/ ? $1 : '' } $a, $b;
        $c1 cmp $c2 || length $b <=> length $a || $a cmp $b;
} @words;
print "@words\n";

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>