save-ebuild-env.sh: refactor the __save_ebuild_env() function
The "save-ebuild-env.sh" unit implements a function by the name of
__save_ebuild_env(), the purpose of which is to dump a subset of the
functions and variables that belong to the current shell's operating
environment. The resulting declarations are used to populate the
"environment.bz2" files that are a component of Portage's internal
database, known as the VDB. This function suffers from several minor
issues, which are described and addressed herewith.

The function calls upon the has() function to determine whether the
--exclude-init-phases option has been specified. However, it expands the
positional parameters as $*, potentially resulting in word splitting and
unwanted pathname expansion. Address this issue by properly expanding
them as "$@", duly rectifying instances of SC2048 and SC2086.

The function consists of 18 invocations of the unset builtin, some of
which are given a large number of parameters. These are neither sorted
nor wrapped in a consistent manner. Further, quoted <newline> characters
are used as a line continuation device, rendering some of the commands
unpleasant to read and difficult to rearrange. Address these issues in a
manner twofold. Firstly, by purposing the REPLY variable as an array
that is used to contain the parameters, thereby allowing for the
arbitrary use of whitespace to separate each. Secondly, by using a
specialised sorting algorithm - described by one of the referenced
commits - to reorder the elements, while rewrapping to 80 columns.

The unset builtin is presently invoked without options. Consequently, it
shall consider each operand as the name of a variable to be unset. In
the event that no variable by the given name exists, it shall instead
attempt to destroy a function by that name. Address this issue by always
specifying the -v option in the case that variables are to be unset, and
the -f option in the case that functions are to be destroyed.

See-also: 93275ac104ad6999f9c4551ceb150c047fca1979
See-also: 44cc0ab2caf13a18cbb99580b91a8afe61ce9cc4
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>