save-ebuild-env.sh: filter out variables that can influence coreutils
As of the time of writing, there are 18 variables that may influence the
behaviour of the utilities comprising the GNU coreutils project. Ensure
that they are filtered out by the __save_ebuild_env() function. Some,
such as 'HOME' and 'TMPDIR', continue to be specified elsewhere. Such is
harmless. I would sooner be thorough in the course of listing variables
that can be directly attributed to a given source.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>