isolated-functions.sh: drop the __assert_sigpipe_ok() function
The __assert_sigpipe_ok() function was introduced in 2010 as a means of
helping to handle the processing of malformed tarballs being distributed
for various sub-projects of GNOME. Among them were the tarballs for
glibmm and evolution-data-server, both of which were found to be the
improper concatenations of two tar archives of the archaic "v7" format.

As things stand now, fifteen years later, I believe it to be a hack that
has run its course. Nowadays, it is conventional for both GNU tar and
bsdtar to produce archives that are of the "pax" format, as defined by
POSIX.1-2001. This is a format unencumbered by the limitations that were
indirectly responsible for the aforesaid defective tarballs. Besides
which, the unpack() function of the "phase-helpers.sh" unit no longer
calls __assert_sigpipe_ok() (since commit 6a3cd41450).

Hence, jettison the __assert_sigpipe_ok() function. Further, jettison
the 'PORTAGE_SIGPIPE_STATUS' environment variable that had hitherto been
defined by the "doebuild.py" unit. The purpose of this variable was to
render the function aware of the numerical value of SIGPIPE. Yet, the
shell requires no assistance in determining this information.

$ read -ra sigs < <(set -o posix; kill -l)
$ for i in "${!sigs[@]}"; do [[ ${sigs[i++]} == PIPE ]] && break; done
$ echo "$i"
13

Link: https://bugs.gentoo.org/309001
Link: https://bugzilla.gnome.org/show_bug.cgi?id=613808#c9
See-also: 6a3cd41450a78422248f10e1272fdd529bd8bc98
See-also: 0ab6f83070c9a32a24454053319c99559c2654ca
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>