isolated-functions.sh: don't open/close/dup file descriptors so frequently
As regards the die() function, refrain from duplicating file descriptor
#2 to file descriptor #1 for each invocation of the functions whose
names comprise the value of the 'EBUILD_DEATH_HOOKS' variable, twice.
Instead, duplicate once and maintain the state of file descriptor #1 for
the duration of the containing loop.

As regards the __elog_base() function, refrain from duplicating file
descriptor #2 to file descriptor #1 twice in the course of printing the
banner message. Instead, do so only once. Further, refrain from opening
and closing a log file for each and every line that is destined to it.
Instead, open the file once and keep it open for the duration of the
containing loop.

As regards the eqawarn(), eerror(), einfo(), ewarn() and elog()
functions, refrain from duplicating file descriptor #2 to file
descriptor #1 for each and every line that is to be printed. Instead,
duplicate once and maintain the state of file descriptor #1 for the
duration of the containing loop.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>