isolated-functions.sh: don't validate the first parameter in __elog_base()
Presently, the __elog_base() helper function validates the first
positional parameter, which is expected to be any of "INFO", "WARN",
"ERROR", "LOG" or "QA". To do so is futile because there are only six
instances in which the function is invoked, all of which hard-code one
of these expected values. Drop the validation code.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>