isolated-functions.sh: mute two instances of SC2185 that are false-positives
Presently, there are two implementations of find0(), one of which serves
to cover situations where the -files0-from option is unavailable,
thereby retaining compatibility with the outdated GitHub CI environment.

Presently, both the preferred implementation and the test that
determines which implementation to use raise spurious SC2185 warnings,
because shellcheck is unaware of the option and its purpose. Mute this
warning code for the entirety of the "isolated-functions.sh" unit.

Link: https://www.shellcheck.net/wiki/SC2185
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>