isolated-functions.sh: specify the -r option to read in __repo_attr()
Presently, the __repo_attr() function acts as a crude parser for the
value of the 'PORTAGE_REPOSITORIES' variable, which contains the
contents of files beneath the "/etc/portage/repos.conf" directory. It
makes 'PORTAGE_REPOSITORIES' the subject of a herestring, and uses the
read builtin to obtain each embedded line. However, if fails to specify
the -r option. Have it do so, duly rectifying an instance of SC2162.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>