isolated-functions.sh: address two cases where shellcheck advises against let
As a matter of style, shellcheck recommends that the (( … )) compound
command be preferred over the let builtin for arithmetic. Presently, it
does so for two instances in which the let builtin is being used.

The first concerns the assert() function, where the builtin is used out
of necessity. Mute the recommendation there.

The second concerns the ebegin() function, where there is no need to use
the builtin. Have it use the (( … )) command instead.

See-also: 6f9971e54dab95c86f0eced52574a586916a5e4b
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>