isolated-functions.sh: render the control flow of __repo_attr() clear
Presently, the way in which the __repo_attr() function is written
renders its control flow obscure. Address this issue by employing a
nicely structured if command and treating the 'appropriate_section'
variable as a quasi-boolean that is evaluated in the arithmetic context.

Further, remove a comment pointing out that bash-3.2 may not handle
extended globs in conditional expressions, if presented in their raw
form. While such is true of 3.2 - and all versions of bash prior to 4.1
- it is irrelevant at this juncture.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>