isolated-functions.sh: drop the RC_DOT_PATTERN variable
Normally, the ebegin() function composes a message comprising the
positional parameters, having been joined in accordance with the state
of the 'IFS' variable, followed by a <space> and three <period>
characters (so as to resemble an ellipsis). However, if the
'RC_DOT_PATTERN' variable is defined as a non-empty string, the function
will instead partially right-pad the composed message with its contents.

$ stty size
24 79
$ ebegin "Got to be starting something" # 35 chars printed
 * Got to be starting something ...

$ RC_DOT_PATTERN=_
$ ebegin "Got to be starting something" # 72 chars printed
 * Got to be starting something_________________________________________

OpenRC does not implement this feature. Nor does gentoo-functions,
though it used to define the 'RC_DOT_PATTERN' variable (despite making
no use of it). I can find no evidence that the feature is being used and
am hard-pressed to believe that anyone would miss it. Get rid of it.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>