isolated-functions.sh: suppress the SC2219 notice throughout
As a matter of style, shellcheck recommends that the (( … )) compound
command be preferred over the let builtin for arithmetic. Presently,
this recommendation is being suppressed for the body of the assert()
function. Instead, suppress it for the "isolated-functions.sh" unit in
its entirety. The reason for this change is that a forthcoming commit
will introduce an additional function that uses the let builtin.

See-also: 91283cf5b59a482061bf4cf440e774a0826c664b
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>