isolated-functions.sh: add and integrate the configparser() function
Presently, the __repo_attr() function acts as a crude parser for the
value of the 'PORTAGE_REPOSITORIES' environment variable, which contains
the contents of files beneath the "/etc/portage/repos.conf" directory.
Its implementation is substandard, to say the least. Firstly, it suffers
from several amateurish bugs. Secondly, it is unable to detect errors in
syntax. Thirdly, it is an incomplete, incorrect implementation of a
parser for the file format that it is supposed to handle.

This commit incorporates a new function by the name of configparser().
It consumes the standard input and - as its name suggests - attempts to
parse it as the "configparser" configuration file format that is native
to python. It has been tested against the sample file provided by the
python documentation and fully supports multiline values and key/value
pairs that are separated by the <colon> character. Each key/value entry
that is encountered shall be printed in the following format:

  "%s\0%s\0%s\0", <section>, <key>, <value>

It is capable of detecting syntax errors. Upon encountering any such
error, a diagnostic message shall be printed to standard error and the
function shall immediately return a non-zero status.

This commit also revises the __repo_attr() function so as to call the
configparser() function and interpret its output.

Fixes: e72b8b90542d1ed2bf53c2b9a0491795c12e63c2
Link: https://docs.python.org/3/library/configparser.html
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>