ebuild.sh: require bash-4.4 and improve __check_bash_version()
As of the time of writing, portage makes use of features that were
introduced by bash-4.4 and, therefore, cannot run correctly with any
older version. However, the __check_bash_version() function does not
currently enforce this requirement.

Address this issue by specifying the minimum required version
independently of the shell compatibility level. The minimum required
version is initially defined as 4.4 but is raised to 5.0 in the event
that the currently effective EAPI requires it (as EAPI 8 does). As is
depicted beneath, the diagnostic messages make this distinction clear.

 * ERROR: sys-apps/portage-9999::gentoo failed (pretend phase):
 *   Portage requires >=bash-4.4 but only 4.3 was found

 * ERROR: sys-apps/portage-9999::gentoo failed (pretend phase):
 *   EAPI 8 requires >=bash-5.0 but only 4.4 was found

Further, apply some minor improvements to the function at large. These
improvements are described herewith.

Remove the comment preceding the function declaration, for it does not
describe its behaviour accurately and is of questionable value.

Clean up the version test by employing an explicit arithmetic context.

Refrain from using the shopt builtin to set the compatibility level.
There is no longer any need to set the compatibility level other than by
assigning to the 'BASH_COMPAT' variable.

Rewrite the remaining commentary so as to employ a more formal tone and
be more informative, without being unduly verbose. Incorporate links to
supporting documentation and associated bug reports, where appropriate.

Attempt to disable the 'patsub_replacement' and 'globskipdots' shell
options on an unconditional basis while simply muting STDERR for the
benefit of bash releases prior to 5.2. As the preceding comment states,
it may become necessary to disable these options on a conditional basis
for a future EAPI. For now, no such logic is required.

See-also: 59b81ecda599e9fbf4084e4595f1fa96c32a9fc1
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>