save-ebuild-env.sh: don't strip PATH
A recent commit added PATH to the list of variables we strip out when
environment saving, but this defeats e.g. llvm-r{1,2}.eclass's mangling
of PATH, showing up when building dev-lang/python[jit]. But it's pretty
common to set PATH somewhere in the ebuild and expect it to persist into
other phases.

Instances of SHELL and TZ could do with checking in ::gentoo to see if
we need to change that too.

Fixes: 1df2556c3aeebc64f64c5e6a2f92d9f4a68eb9da
Signed-off-by: Sam James <sam@gentoo.org>