Don't strip SHELL while saving and restoring environments
The 'SHELL' variable is set by whatever is responsible for handling the
procedure of logging in, be it login(1), sshd(8) or a display manager.
Its value is taken from the user's password entry and specifies their
choice of login shell. Additionally, bash(1) sets the variable, albeit
only in the (irregular) case that it is missing in its environment at
the time of execution. Ideally, it should not be set by other programs,
nor should any other meaning or purpose be ascribed to it.

Unfortunately, some packages define and export the 'SHELL' variable
during the src_configure phase. Among them are the following:

- app-admin/entr
- dev-lang/spidermonkey
- mail-client/thunderbird
- www-client/firefox
- www-client/seamonkey

While there is a reason for their doing so, it is not immediately clear
as to whether there is an actual need for the variable to persist
between ebuild phases. That is, it may be that the variable only needs
to be present in the environment of the configure program. Still, until
such time as this can be established with absolute certainty, it would
be prudent to avoid the risk of incurring a regression.

For now, refrain from filtering out the 'SHELL' variable in both the
__save_ebuild_env() function of the "save-ebuild-env.sh" unit, and the
__filter_readonly_variables() function of the "phase-functions.sh" unit.

See-also: 8a607ed23beb4995627f9c4e2c7580469ace56bb
See-also: 1df2556c3aeebc64f64c5e6a2f92d9f4a68eb9da
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>